chore: format

+++ clox/object.c
@@ -31,9 +31,10 @@ static ObjString* allocateString(char* chars, int length, uint32_t hash) {
 static uint32_t hashString(const char* key, int length) {
     uint32_t hash = 2166136261u;
     for (int i = 0; i < length; i++) {
-        hash ^= key[i];
+        hash ^= (uint8_t)key[i];
         hash *= 16777619;
     }
+    
     return hash;
 }
